# backend/src/main/java/com/aivle/backend/journey/MarketingReportJourneyController.java
package com.aivle.backend.journey;import com.aivle.backend.common.response.ApiResponse;import com.aivle.backend.common.security.CurrentUserProvider;import jakarta.servlet.http.HttpServletRequest;import jakarta.validation.Valid;import lombok.RequiredArgsConstructor;import org.springframework.web.bind.annotation.*;
@RestController @RequestMapping("/api/v2/projects/{projectId}") @RequiredArgsConstructor
public class MarketingReportJourneyController{private final MarketingReportJourneyService service;private final CurrentUserProvider currentUser;
 @PostMapping("/marketing-generations")public ApiResponse<MarketingReportJourneyService.WorkspaceView> generate(@PathVariable Long projectId,HttpServletRequest request){return ApiResponse.success(service.generate(user(),projectId),id(request));}
 @GetMapping("/marketing-workspace")public ApiResponse<MarketingReportJourneyService.WorkspaceView> workspace(@PathVariable Long projectId,HttpServletRequest request){return ApiResponse.success(service.workspace(user(),projectId),id(request));}
 @PutMapping("/marketing-assets/{assetId}/select")public ApiResponse<MarketingReportJourneyService.WorkspaceView> select(@PathVariable Long projectId,@PathVariable Long assetId,HttpServletRequest request){return ApiResponse.success(service.selectAsset(user(),projectId,assetId),id(request));}
 @PostMapping("/marketing-comparisons")public ApiResponse<MarketingReportJourneyService.ComparisonView> compare(@PathVariable Long projectId,HttpServletRequest request){return ApiResponse.success(service.compare(user(),projectId),id(request));}
 @GetMapping("/marketing-comparisons/current")public ApiResponse<MarketingReportJourneyService.ComparisonView> currentComparison(@PathVariable Long projectId,HttpServletRequest request){return ApiResponse.success(service.currentComparison(user(),projectId),id(request));}
 @PostMapping("/final-reports")public ApiResponse<MarketingReportJourneyService.ReportView> report(@PathVariable Long projectId,HttpServletRequest request){return ApiResponse.success(service.report(user(),projectId),id(request));}
 @GetMapping("/final-reports/current")public ApiResponse<MarketingReportJourneyService.ReportView> currentReport(@PathVariable Long projectId,HttpServletRequest request){return ApiResponse.success(service.currentReport(user(),projectId),id(request));}
 @PutMapping("/final-reports/{reportId}/decision")public ApiResponse<MarketingReportJourneyService.ReportView> decision(@PathVariable Long projectId,@PathVariable Long reportId,@Valid @RequestBody MarketingReportJourneyService.DecisionRequest body,HttpServletRequest request){return ApiResponse.success(service.decide(user(),projectId,reportId,body),id(request));}
 private Long user(){return currentUser.currentUserId();}private String id(HttpServletRequest r){return r.getHeader("X-Request-Id");}}