# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Business Validation Platform Document API
  version: 0.12.0-phase3-taskrun
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >
    Canonical authentication, document upload, asynchronous processing,
    structured-plan mutation, and explicit confirmation contract.
servers:
  - url: /api/v1
...
security:
  - platformAuth: []
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List the current user's active projects
      operationId: listProjects
      x-implementation-phase: Phase 5
      responses:
        '200':
          description: Projects ordered by most recently updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectListApiResponse'}
        '401': {$ref: '#/components/responses/StandardError'}
    post:
      tags: [Projects]
      summary: Create a business validation project
      operationId: createProject
      x-implementation-phase: Phase 5
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateProjectRequest'}
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectDetailApiResponse'}
        '400': {$ref: '#/components/responses/StandardError'}
        '401': {$ref: '#/components/responses/StandardError'}
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: Get a project owned by the current user
      operationId: getProject
      x-implementation-phase: Phase 5
      responses:
        '200':
          description: Owned project detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectDetailApiResponse'}
        '401': {$ref: '#/components/responses/StandardError'}
        '404':
          description: Project does not exist or is not visible to the current user
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
...
  /projects/{projectId}/persona-recommendations:
    post:
      tags: [Personas]
      summary: Start a persona recommendation and customer-validation planning job
      description: >
        Compares the latest complete confirmed plan and its persisted feasibility
        assessment with the aggregate catalog. The result is a hypothesis, not
        observed customer evidence, purchase probability, or completed validation.
      operationId: startPersonaRecommendation
      x-implementation-phase: Phase 10
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '202':
          description: Job accepted, or the completed idempotent result returned
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PersonaStartApiResponse'}
        '404': {$ref: '#/components/responses/StandardError'}
        '409': {$ref: '#/components/responses/StandardError'}
        '422': {$ref: '#/components/responses/StandardError'}
  /projects/{projectId}/persona-recommendations/latest:
    get:
      tags: [Personas]
      summary: Get the latest persona recommendation and validation plan
      operationId: getLatestPersonaRecommendation
      x-implementation-phase: Phase 10
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Latest owner-visible recommendation
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PersonaRecommendationApiResponse'}
        '404': {$ref: '#/components/responses/StandardError'}
...
  /api/v2/projects/{projectId}/task-runs/{taskRunId}:
    get:
      tags: [Jobs]
      servers:
        - url: /
      summary: Read an owner-scoped Target TaskRun
      operationId: getTaskRunV2
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {$ref: '#/components/parameters/TaskRunId'}
      responses:
        '200':
          description: Current state, including terminal failure states
          headers:
            X-Correlation-Id:
              description: Request correlation identifier
              schema: {type: string, maxLength: 128}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskRunV2Response'}
        '404': {$ref: '#/components/responses/TaskRunV2Error'}
  /api/v2/projects/{projectId}/task-runs/{taskRunId}/retry:
    post:
      tags: [Jobs]
      servers:
        - url: /
      summary: Queue a new attempt on the same retryable TaskRun
      operationId: retryTaskRunV2
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {$ref: '#/components/parameters/TaskRunId'}
        - {$ref: '#/components/parameters/TaskCommandIdempotencyKey'}
      responses:
        '202':
          description: Retry accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskRunV2Response'}
        '404': {$ref: '#/components/responses/TaskRunV2Error'}
        '409': {$ref: '#/components/responses/TaskRunV2Error'}
  /api/v2/projects/{projectId}/task-runs/{taskRunId}/cancel:
    post:
      tags: [Jobs]
      servers:
        - url: /
      summary: Idempotently request TaskRun cancellation
      operationId: cancelTaskRunV2
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {$ref: '#/components/parameters/TaskRunId'}
      responses:
        '200':
          description: Current TaskRun state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskRunV2Response'}
        '404': {$ref: '#/components/responses/TaskRunV2Error'}
components:
  securitySchemes:
    platformAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Phase 3 access JWT. The development X-User-Id adapter is profile-bound
        and is deliberately excluded from the public contract.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {type: integer, format: int64}
    TaskRunId:
      name: taskRunId
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 64}
...
    TaskCommandIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required command replay key scoped to owner, project, method and normalized path.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
...
  responses:
    TaskRunV2Error:
      description: Public v2 safe error envelope
      content:
        application/json:
          schema: {$ref: '#/components/schemas/TaskRunV2ErrorResponse'}
    StandardError:
      description: Standard API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
  schemas:
    TaskRunV2ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId, taskRunId, details]
          properties:
            code: {type: string}
            message: {type: string}
            correlationId: {type: string, maxLength: 128}
            taskRunId: {type: [string, 'null'], maxLength: 64}
            details:
              type: array
              items:
                type: object
                required: [field, reason, resourceType, resourceId]
                properties:
                  field: {type: [string, 'null']}
                  reason: {type: string}
                  resourceType: {type: [string, 'null']}
                  resourceId: {type: [string, 'null']}
    TaskRunV2Response:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/TaskRunPublicViewV2'}
        meta:
          type: object
          required: [correlationId]
          properties:
            correlationId: {type: string, maxLength: 128}
    TaskRunPublicViewV2:
      type: object
      required: [id, taskType, subject, state, retryable, cancelable, correlationId, createdAt]
      properties:
        id: {type: string, maxLength: 64}
        taskType: {type: string}
        subject:
          type: object
          required: [type, id]
          properties:
            type: {type: string, maxLength: 80}
            id: {type: string, maxLength: 64}
        state:
          type: string
          enum: [QUEUED, READY, RUNNING, SUCCEEDED, FAILED, CANCELLED, TIMED_OUT]
        retryable: {type: boolean}
        cancelable: {type: boolean}
        correlationId: {type: string, maxLength: 128}
        createdAt: {type: string, format: date-time}
        startedAt: {type: [string, 'null'], format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
        errorSummary:
          type: [object, 'null']
          properties:
            code: {type: string}
        resultResource:
          type: [object, 'null']
          properties:
            type: {type: string}
            id: {type: string}
    Meta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: {type: string}
        timestamp: {type: string, format: date-time}
    ApiError:
      type: object
      required: [code, message, fieldErrors, retryable]
      properties:
        code: {type: string}
        message: {type: string}
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}
        retryable: {type: boolean}
    ErrorApiResponse:
      type: object
      required: [success, error, meta]
      properties:
        success: {type: boolean, const: false}
        error: {$ref: '#/components/schemas/ApiError'}
        meta: {$ref: '#/components/schemas/Meta'}
...
    CreateProjectRequest:
      type: object
      required: [title]
      properties:
        title: {type: string, minLength: 1, maxLength: 150}
        description:
          type: [string, 'null']
          maxLength: 10000
        industryCategory:
          type: [string, 'null']
          maxLength: 100
    ProjectStatus:
      type: string
      enum: [DRAFT, ACTIVE, PAUSED, COMPLETED, ARCHIVED]
    ProjectStage:
      type: string
      enum:
        - DOCUMENT
        - STRUCTURING
        - LEGAL_REVIEW
        - FEASIBILITY
        - FINANCIAL
        - PERSONA_CONFIGURATION
        - PANEL_SURVEY
        - PANEL_DISCUSSION
        - REPORT
        - MARKETING
        - COMPLETED
    ProjectSummary:
      type: object
      required: [id, title, stage, status, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64}
        title: {type: string}
        industryCategory: {type: [string, 'null']}
        stage: {$ref: '#/components/schemas/ProjectStage'}
        status: {$ref: '#/components/schemas/ProjectStatus'}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/ProjectSummary'
        - type: object
          required: [ownerId, version]
          properties:
            ownerId: {type: integer, format: int64}
            description: {type: [string, 'null']}
            startedAt: {type: [string, 'null'], format: date-time}
            completedAt: {type: [string, 'null'], format: date-time}
            version: {type: integer, format: int64, minimum: 0}
...
    JobStatus:
      type: string
      enum:
        - NOT_STARTED
        - QUEUED
        - RUNNING
        - SUCCEEDED
        - PARTIAL
        - FAILED
        - CANCELED
...
    ProjectListApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data:
          type: array
          items: {$ref: '#/components/schemas/ProjectSummary'}
        meta: {$ref: '#/components/schemas/Meta'}
    ProjectDetailApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/ProjectDetail'}
        meta: {$ref: '#/components/schemas/Meta'}
...
    PersonaStart:
      type: object
      required: [projectId, jobId, status, structuredPlanId,
        feasibilityAssessmentId, catalogVersion]
      properties:
        projectId: {type: integer, format: int64}
        recommendationId: {type: [integer, 'null'], format: int64}
        jobId: {type: integer, format: int64}
        status: {$ref: '#/components/schemas/JobStatus'}
        structuredPlanId: {type: integer, format: int64}
        feasibilityAssessmentId: {type: integer, format: int64}
        catalogVersion: {type: string, const: persona-catalog-v1}
    PersonaRecommendation:
      type: object
      required: [recommendationId, projectId, analysisJobId, structuredPlanId,
        feasibilityAssessmentId, status, confidence, summary, disclaimer,
        provider, modelName, promptVersion, catalogVersion, completedAt,
        items, hypotheses, validationPlans, linkedFeasibilityTasks]
      properties:
        recommendationId: {type: integer, format: int64}
        projectId: {type: integer, format: int64}
        analysisJobId: {type: integer, format: int64}
        structuredPlanId: {type: integer, format: int64}
        feasibilityAssessmentId: {type: integer, format: int64}
        status: {type: string, enum: [COMPLETED, NEEDS_VALIDATION]}
        primaryPersonaCode: {type: [string, 'null']}
        secondaryPersonaCode: {type: [string, 'null']}
        confidence: {type: string, enum: [LOW, MEDIUM, HIGH]}
        summary: {type: string}
        disclaimer: {type: string}
        provider: {type: string}
        modelName: {type: string}
        promptVersion: {type: string, const: persona-recommendation-v1}
        catalogVersion: {type: string, const: persona-catalog-v1}
        completedAt: {type: string, format: date-time}
        items: {type: array, minItems: 2, items: {type: object}}
        hypotheses: {type: array, items: {type: object}}
        validationPlans: {type: array, items: {type: object}}
        linkedFeasibilityTasks: {type: array, items: {type: object}}
...
    PersonaStartApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/PersonaStart'}
        meta: {$ref: '#/components/schemas/Meta'}
    PersonaRecommendationApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/PersonaRecommendation'}
        meta: {$ref: '#/components/schemas/Meta'}